requirements.txt
fastapi
uvicorn
langchain-community
faiss-cpu
sentence-transformers
google-generativeai